- Added Struct_1D530 to allow better management of the map data. This should allow build an extra area viewer, to allow debugging of scripting issues. - Renaming

+++ trunk/engine/seg040.cs
@@ -4,7 +4,7 @@ namespace engine
 {
     class seg040
     {
-        internal static void load_dax(ref DaxBlock mem_ptr, byte mask_colour, byte masked, byte block_id, string fileName)
+        internal static void load_dax(ref DaxBlock mem_ptr, byte mask_colour, byte masked, int block_id, string fileName)
         {
             DaxBlock dax_ptr;
             byte[] pic_data;
